Move package-generator into generate-packages
Unsplitting the code is way more readable.

Also slightly improve logging in calculate-versions.
